Move network module to controller
