--- framework/src/modules/network/components/storage/sql/network_info/get.sql
@@ -1,21 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-SELECT * FROM network_info
-
-${parsedFilters:raw}
-
-${parsedSort:raw}
-
-LIMIT ${limit} OFFSET ${offset}
