--- framework/src/modules/network/components/storage/sql/network_info/upsert.sql
@@ -1,20 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-
-INSERT INTO "network_info" ("key", "value")
-VALUES(${key},${value})
-ON CONFLICT ON CONSTRAINT network_info_pkey
-DO
-	UPDATE SET "value" = ${value};
